Add partial skeletons for the guide and 404
These go together because the method for determining which guide it is
takes an unusual form. Instead of having a guide route for every person,
there's one catch all guide page which tries to parse out a guide name
and then routes to it. If it doesn't find that guide's name, it'll just
redirect to the 404.html page.